# static/swagger.yaml
openapi: 3.0.0
info:
  title: API de Verificación de Rostros y Usuarios
  description: API para la verificación de rostros, inicio de sesión y registro de usuarios.
  version: 1.0.1
servers:
  - url: http://localhost:5000
    description: Servidor local
  - url: https://living-improved-ocelot.ngrok-free.app/
    description: Servidor en linea
...
paths:
...
  /api/class/delete:
    delete:
      summary: Delete a class
      description: Deletes a class based on the provided class_id.
      tags:
        - Clases
      parameters:
        - in: query
          name: class_id
          required: true
          schema:
            type: string
          description: ID of the class to delete.
      responses:
        '200':
          description: Class deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  status_code:
                    type: integer
        '400':
          description: Missing class_id parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
  /api/class/add-student:
    post:
      summary: Add a student to a class
      description: Registers a student into a class based on student_id and class_id.
      tags:
        - Clases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                student_id:
                  type: integer
                  description: ID of the student.
                class_id:
                  type: integer
                  description: ID of the class.
              required:
                - student_id
                - class_id
      responses:
        '201':
          description: Student added to class successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  status_code:
                    type: integer
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
  /api/class/delete-student:
    delete:
      summary: Delete a student from a class
      description: Deletes a student from a class based on student_id and class_id.
      tags:
        - Clases
      parameters:
        - in: query
          name: class_id
          required: true
          schema:
            type: string
          description: ID of the class.
        - in: query
          name: student_id
          required: true
          schema:
            type: string
          description: ID of the student.
      responses:
        '200':
          description: Student deleted from class successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  status_code:
                    type: integer
        '400':
          description: Missing class_id or student_id parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
...
  /api/class/exams:
    get:
      summary: Retrieve exams for a class
      description: Retrieves all exams for a class based on class_id.
      tags:
          - Clases
      parameters:
        - in: query
          name: class_id
          required: true
          schema:
            type: string
          description: ID of the class.
      responses:
        '200':
          description: Exams retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  status_code:
                    type: integer
        '400':
          description: Missing class_id parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
  /api/class/students:
    get:
      summary: Retrieve students for a class
      description: Retrieves all students for a class based on class_id.
      tags:
        - Clases
      parameters:
        - in: query
          name: class_id
          required: true
          schema:
            type: string
          description: ID of the class.
      responses:
        '200':
          description: Students retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  status_code:
                    type: integer
        '400':
          description: Missing class_id parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
...
  /api/teacher/classes:
    get:
      summary: Retrieve classes for a teacher
      description: Retrieves all classes for a teacher based on teacher_id.
      tags:
        - Profesores
      parameters:
        - in: query
          name: teacher_id
          required: true
          schema:
            type: string
          description: ID of the teacher.
      responses:
        '200':
          description: Classes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  status_code:
                    type: integer
        '400':
          description: Missing teacher_id parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
...
  /api/exam/create:
    post:
      summary: Create a new exam
      description: Creates a new exam with the provided details.
      tags:
        - Exámenes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                exam_name:
                  type: string
                  description: Name of the exam.
                class_id:
                  type: integer
                  description: ID of the class.
                date:
                  type: string
                  description: Date of the exam.
                class_room:
                  type: string
                  description: Class room.
                hour:
                  type: string
                  description: Hour of the exam.
              required:
                - exam_name
                - class_id
      responses:
        '201':
          description: Exam created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  status_code:
                    type: integer
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
...